requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.6.0
Flask-WTF==1.2.2
python-dotenv==1.2.1
Werkzeug==3.1.6
SQLAlchemy==2.0.46
alembic==1.18.4
eventlet==0.40.4
Pillow==12.1.1
qrcode==8.2
gunicorn==25.1.0
email-validator==2.3.0
MarkupSafe==3.0.3
